WIP: debugtext moved to debug. Debug Folder created. TODO: need to separate debug and debugtext. onvalidate fix needed, relative pos, etc
