+++ com.unity.cinemachine/Runtime/Debug.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad006de08e6604a8bb48bc0e7db399a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
